refactor: Rename generator module to maker
To be more precise to what it really does...
